Changing gameplay

+++ Assets/Zoomania/Scripts/Level_System/Buildings/Improvement_Level_New.cs
@@ -8,6 +8,6 @@ public class Improvement_Level_New
 {
 	[field: SerializeField] public Sprite View { get; set; }
 	[field: SerializeField] public int CurrentLevelNumber { get; set; }
-	[field: SerializeField] public int EffectValue { get; set; }
+	[field: SerializeField] public float EffectValue { get; set; }
 	[field: SerializeField] public int MoneyForUpgrade { get; set; }
 }
