+++ Assets/Zoomania/Scripts/Timer/Timer.cs
@@ -30,7 +30,7 @@ public class Timer
 		Pause = pause;
 	}
 
-	public void SetMaxTimeAndReset(int maxtime)
+	public void SetMaxTimeAndReset(float maxtime)
 	{
 		TimerIsEnd = false;
 		CurrentTime = 0;
